pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dj-agent"
version = "0.2.0"
description = "AI-powered Rekordbox library enrichment tool"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pyrekordbox>=0.4.0",
    "mutagen>=1.47.0",
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "pyloudnorm>=0.1.0",
    "soundfile>=0.12.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "pydub>=0.25.1",
    "rapidfuzz>=3.0.0",
]

[project.optional-dependencies]
fingerprint = ["pyacoustid>=1.3.0"]
similarity = ["faiss-cpu>=1.7.0"]
mood = ["essentia-tensorflow>=2.1b6", "laion-clap>=1.1.4"]
stems = ["demucs>=4.0"]
normalize = ["ffmpeg-normalize>=1.28.0"]
master = ["pedalboard>=0.9.0"]
gui = ["gradio>=4.0"]
metadata = ["musicbrainzngs>=0.7.1", "pylast>=5.0", "discogs-client>=2.3"]
quality = ["flac-detective"]
phrases = ["madmom"]
all = [
    "dj-agent[fingerprint,similarity,mood,stems,normalize,master,metadata,quality,phrases]",
]
dev = ["pytest>=7.4", "pytest-cov>=4.1"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
